Restructure utils & add comments
